git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/cortical_surface/cortical_surface-gpl/branches/bug_fix@79620 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ brainvisa/toolboxes/cortical_surface/processes/anatomy/tools/InsularPoleProjection.py
@@ -136,6 +136,6 @@ def execution( self, context ):
     cingular_tex_clean, cing_tex_boundary = textureTls.textureTopologicalCorrection(mesh, tex[0].arraydata(), tmp_tex_value)
     cingular_tex_clean[np.where(cingular_tex_clean == tmp_tex_value)[0]] = 180
     tex_out = aims.TimeTexture_S16()
-    tex_out[0].assign(cingular_tex_clean)
+    tex_out[0].assign(cingular_tex_clean.astype(np.int16))
     ws.write(tex_out, self.pole.fullPath())
     context.write('... Done')
